giscanner: rename ast.py to gi_ast.py
Having a module share the same name as a built-in module is never
a good idea, and can manifest in incorrect imports when using pyenv.

In this specific case, the newly-imported pycparser imports inspect,
which in turn tries to import ast, expecting the builtin-module to
be imported but getting our module instead.