Add "(static-method)" annotation
static-method tag will force a function to be a static method.

For a non-gobject function like `void itemlist_selection_changed (Item item)`,
the function was supposed to be a static method of a boxed type ItemList
as `ItemList.selection_changed()`.

However, since the 1st parameter of the function is an Item type, the
transformer will smart guessing it into a method of
`Item.ist_selection_changed()` or something.

A `static-method` tag will prevent "the smart guessing into method" from
happening.